+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Version 1.2.2 (unreleased)
 - bump certifi 2023.7.22 -> 2024.7.4 (from PR #233)
 - bump zipp 3.3.1 -> 3.19.1 (from PR #234)
 - bump ipython 7.16.1 -> 8.10.0 (from PR #239)
+- add FIFO lot sorting for splits in accounts (from PR #237, @akjore)
 
 Version 1.2.1 (2024-07-14)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~